move actor-component to components
